refactoring balance game structure
